requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
matplotlib>=3.7.0
tensorboard>=2.13.0
pytest>=7.3.0
gradio>=4.0.0
noisereduce>=3.0.0
speechbrain>=1.0.0
soundfile>=0.12.0
requests>=2.31.0
tqdm>=4.65.0
